Rename concat.rb to be more consistent with the class name
