+++ lib/rake-pipeline/filters.rb
@@ -1,2 +1,2 @@
-require "rake-pipeline/filters/concat"
+require "rake-pipeline/filters/concat_filter"
 require "rake-pipeline/filters/ordering_concat_filter"
