+++ lib/rake-pipeline/filters/concat_filter.rb
